third_party/blink/web_tests/external/wpt/html/dom/documents/resource-metadata-management/document-cookie-expected.txt

This is a testharness.js-based test.
PASS document has no cookie
PASS document.cookie
PASS document.cookie 1
FAIL document.cookie 2 assert_equals: expected "" but got "b=A"
PASS getting cookie for a cookie-averse document returns empty string, setting does nothing
Harness: the test ran to completion.
